+++ services/workflows-service/src/events/workflow-state-changed-webhook-caller.ts
@@ -49,15 +49,11 @@ export class WorkflowStateChangedWebhookCaller {
       'workflow.state.changed',
     );
 
-    const customer = await this.customerService.getByProjectId(
-      // @ts-expect-error - error from Prisma types fix
-      data.runtimeData.projectId,
-      {
-        select: {
-          authenticationConfiguration: true,
-        },
+    const customer = await this.customerService.getByProjectId(data.runtimeData.projectId, {
+      select: {
+        authenticationConfiguration: true,
       },
-    );
+    });
 
     const { webhookSharedSecret } =
       customer.authenticationConfiguration as TAuthenticationConfiguration;