+++ apps/backoffice-v2/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - Updated dependencies
   - @ballerine/ui@0.3.16
+- Updated dependencies [d5cca1e8]
+- Updated dependencies [f4b5c0df]
+  - @ballerine/workflow-browser-sdk@0.5.23
+  - @ballerine/workflow-node-sdk@0.5.23
+  - @ballerine/common@0.7.24
 
 ## 0.5.25